Rename snr module to snratio
Having both a snr module and a snr() function is illegal in Python,
therefore the snr module has been renamed to snratio.
