+++ agenet/__init__.py
@@ -2,7 +2,7 @@
 
 
 from agenet.av_age import av_age_fn
-from agenet.snr import snr, snr_th
+from agenet.snratio import snr, snr_th
 from agenet.bler import blercal, blercal_th
 from agenet.maincom import main, run_main
 from agenet.printplot import generate_table, plot, plot_generate, printage
