+++ test/test_snratio.py
@@ -29,7 +29,7 @@ def test_snr_th():
 
 def test_command_line_arguments():
     """Test the command-line arguments for the snr module."""
-    script_path = os.path.abspath("agenet/snr.py")
+    script_path = os.path.abspath("agenet/snratio.py")
     # Use the value 6000000000 instead of 6*(10**9) or enclose the expression in quotes
     command = f"python {script_path} -N0 0.1 -d 100 -P 1 -fr 6000000000"
     result = subprocess.run(command, shell=True, capture_output=True, text=True)
